Fix the code after API changes in dependencies
The project has loose depedencies, so any change to the API of the used
libraries can easily break the execution of the code. Fix it by:

First, the latest dataclasses_struct requires '@dcs.dataclass_struct'
instead of used '@dcs.dataclass'. Also, we need to be explicit about
the sizing of the fields - because by default it uses native, which
most of the time is wrong while doing the cross-compilation on
different machines. Byteorder default for RISC-V is little endian.
It might be considered to be read from the elf and choosed dynamically
in the future.

Next, the lief library in the latest version is having a bit different
API - align the codebase to be able to read the header correctly.

Last but not least, to avoid such problems in the future, freeze the
dependencies to exact version. That means we don't end up with such
incompatibilities when the dependencies release new versions anymore.
Bump of the underlaying packages version should be explicit, only
after checking if that works.

Signed-off-by: Michał Michalik <michal.michalik.priv@gmail.com>